+++ include/gtirb_pprinter/AuxDataSchema.hpp
@@ -106,12 +106,6 @@ struct BinaryType {
   typedef std::vector<std::string> Type;
 };
 
-/// \brief Auxiliary data describing architecture information
-struct ArchInfo {
-  static constexpr const char* Name = "archInfo";
-  typedef std::map<std::string, std::string> Type;
-};
-
 /// \brief Auxiliary data representing the export table of a PE file.
 struct ExportEntries {
   static constexpr const char* Name = "peExportEntries";