+++ src/gtirb_pprinter/AuxDataUtils.cpp
@@ -111,10 +111,6 @@ std::vector<std::string> getBinaryType(const gtirb::Module& Module) {
   return util::getOrDefault<gtirb::schema::BinaryType>(Module);
 }
 
-std::map<std::string, std::string> getArchInfo(const gtirb::Module& Module) {
-  return util::getOrDefault<gtirb::schema::ArchInfo>(Module);
-}
-
 std::map<gtirb::UUID, gtirb::UUID>
 getSymbolForwarding(const gtirb::Module& Module) {
   return util::getOrDefault<gtirb::schema::SymbolForwarding>(Module);